requirements.txt
wheel
numpy
pygsheets
python-dotenv
Flask
Flask-Admin
peewee
flask_assets
flask_cors
gunicorn
connexion[swagger-ui]
deepdiff
marshmallow_peewee